car/requirements_emulator.txt
nats-py==v2.6.0
nkeys
